Updating Transcription.py
Changing the ._copy() method as well as the .copy() method of the Tier class: If a new Tier is copied and added to a Transcription, each Segment gets copied and added as well as is assigned a new parent Segment depending on whether is has a parent and whether the copied Tier has a different parent than the original Tier object. Also; changing some minor, insignificant things like a forgotten 'self'.