Updating toElan.py
Including a keyword in args** when calling toElan to save a Transcription. This keyword is used to control, whether segments will be renamed (their annotation id) before the Transcription is exported to an .eaf file. By default, segments will be renamed using the letter 'a'. Using a falsy value (except None) for the keyword prevents the function from automatically renaming segments.